Move some blockchain specific stuff out of bitcoinrpc.cpp into rpcblockchain.cpp
