+++ paycoin.pro
@@ -228,6 +228,7 @@ SOURCES += src/qt/bitcoin.cpp src/qt/bitcoingui.cpp \
     src/qt/transactionview.cpp \
     src/qt/walletmodel.cpp \
     src/bitcoinrpc.cpp \
+    src/rpcblockchain.cpp \
     src/rpcdump.cpp \
     src/rpcnet.cpp \
     src/rpcwallet.cpp \
