+++ src/bitcoinrpc.h
@@ -26,6 +26,9 @@ json_spirit::Object JSONRPCError(int code, const std::string& message);
 /** Convert parameter values for RPC call from strings to command-specific JSON objects. */
 json_spirit::Array RPCConvertValues(const std::string &strMethod, const std::vector<std::string> &strParams);
 
+/** Convert unsigned integer/s to hex string */
+std::string HexBits(unsigned int nBits);
+
 typedef json_spirit::Value(*rpcfn_type)(const json_spirit::Array& params, bool fHelp);
 
 class CRPCCommand
@@ -61,14 +64,30 @@ public:
 extern const CRPCTable tableRPC;
 extern CReserveKey* pMiningKey;
 
+// In rpcblockchain.cpp
 extern double GetDifficulty(const CBlockIndex* blockindex = NULL);
+extern json_spirit::Value getblockcount(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value getdifficulty(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value getblock(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value getblockhash(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value settxfee(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value getcheckpoint(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value getrawmempool(const json_spirit::Array& params, bool fHelp);
+// In rpcdump.cpp
+extern json_spirit::Value dumpprivkey(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value importprivkey(const json_spirit::Array& params, bool fHelp);
 // In rpcnet.cpp
 extern json_spirit::Value getconnectioncount(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value getpeerinfo(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value sendalert(const json_spirit::Array& params, bool fHelp);
-// In rpcdump.cpp
-extern json_spirit::Value dumpprivkey(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value importprivkey(const json_spirit::Array& params, bool fHelp);
+// In rpcrawtransaction.cpp
+extern void ScriptPubKeyToJSON(const CScript& scriptPubKey, json_spirit::Object& out);
+extern json_spirit::Value getrawtransaction(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value listunspent(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value createrawtransaction(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value decoderawtransaction(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value signrawtransaction(const json_spirit::Array& params, bool fHelp);
+extern json_spirit::Value sendrawtransaction(const json_spirit::Array& params, bool fHelp);
 // In rpcwallet.cpp
 extern int64 nWalletUnlockTime;
 extern int64 AmountFromValue(const json_spirit::Value& value);
@@ -112,14 +131,6 @@ extern json_spirit::Value repairwallet(const json_spirit::Array& params, bool fH
 extern json_spirit::Value makekeypair(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value clearorphans(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value listmicroprimedata(const json_spirit::Array& params, bool fHelp);
-// In rpcrawtransaction.cpp
-extern void ScriptPubKeyToJSON(const CScript& scriptPubKey, json_spirit::Object& out);
-extern json_spirit::Value getrawtransaction(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value listunspent(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value createrawtransaction(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value decoderawtransaction(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value signrawtransaction(const json_spirit::Array& params, bool fHelp);
-extern json_spirit::Value sendrawtransaction(const json_spirit::Array& params, bool fHelp);
 // In scrapesdb.cpp
 extern json_spirit::Value setscrapeaddress(const json_spirit::Array& params, bool fHelp);
 extern json_spirit::Value getscrapeaddress(const json_spirit::Array& params, bool fHelp);
