+++ src/makefile.linux-mingw
@@ -60,6 +60,7 @@ OBJS= \
     obj/protocol.o \
     obj/kernelrecord.o \
     obj/bitcoinrpc.o \
+    obj/rpcblockchain.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/rpcwallet.o \
